refactor: extract shared optTakesValueEntries to extractor module
Deduplicate identical optTakesValueEntries function from bash.ts and
zsh.ts into the shared extractor.ts, since both generators produce
the same case-branch syntax for opt-takes-value detection.
